FastAPI/todo-mongo/requirements.txt
fastapi
uvicorn
python-dotenv
motor

